CLISApp-backend/requirements.txt
# CLISApp Backend Dependencies
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Geospatial Data Processing
xarray>=2023.12.0
rasterio>=1.3.9
geopandas>=0.14.1
fiona>=1.9.6
pyproj>=3.6.1
shapely>=2.0.2
numpy>=1.24.0  # Array processing
scipy>=1.11.0  # Interpolation and scientific computing

# NetCDF Support
netcdf4==1.6.5
h5netcdf==1.3.0
cfgrib>=0.9.14

# HDF File Support (for GRIB/HDF data)
pyhdf>=0.10.5  # HDF4 file reading
h5py>=3.9.0  # HDF5 support

# Data Download & HTTP
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
tenacity>=8.2.3  # Retry logic for API calls

# Image Processing for Tiles
Pillow>=10.2.0
matplotlib>=3.8.2
colorcet>=3.0.1

# Caching
redis>=5.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Environment & Configuration
python-dotenv==1.0.0
